Allow for HRD on rigs where more than one mode maps to Tranceiver::MODE.
Some rigs  like the FT-857(D) have  multiple modes that only  have one
representation in the Transceiver class,  e.g. FM, FM(N) and WFM.  The
HRDTransceiver class  now allows  multiple modes to  be mapped  to one
Transceiver::MODE, the  last match  in the HRD  mode dropdown  is used
when  mapping  a Tranceiver::MODE  to  a  HRD  mode setting  which  is
normally the most suitable mode.

git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@4135 ab8295b8-cf94-4d9e-aec4-7959e3be5d79