JTSDK-QT Upgrade (major changes)
*UPGRADE INSTRUCTIONS*

JTSDK-DOC should be used to perform the upgrade, as the script uses
a combination Unix & Windows commands, and the archive files are
compressed with 7z which may or may not be installed on the users
system.

Open JTSKD-DOC ( C:\JTSDK-DOC\jtsdk-docenv.bat )
At the prompt, update from SVN .. svn update
At the prompt, run update ....... update

That is all. ( install-scripts.bat ) should copy the new scripts
and install NSIS + Pkg-Config to the correct location(s).

Please report any issues you experience with the upgrade process
or functionality issues.

*MAJOR CHANGES*

* ADDED Pkg-Config-Lite v0.28:
This solves an issue during the Cmake Build process for WSJT-X.
Pkg-Config (which Findhamlib.cmake relies on) was not being found
and causing problems with Hamlib3 library locations. Thus, several
users were having trouble with Hamlib post build. The root of the
problem points to the missing pkg-config package.

* ADDED NSIS v0.3a2 Installer Software:
WSJT-X has a target to build a Windows Installer and uses NSIS as
the tool. We can now build the Windows Installer with a simple cmake
invocation or via the build script:

Build Script ... build wsjtx package
Command-line ... cmake --build . --target package
Output File: wsjtx-1.4.0-win32.exe

* Default File Locations:
This should not have been in the build script at all, as it is a 
CmakeLists.txt option which is defaulted to on. Changing it too
"off" was causing problems with later version of JT-Alerts-X, not to
mention, the implementation was wrong to begin with. If a user needs
to change Cmake config Options, it should be done using the cmake-gui, 
QtCreator or manually editing / configuring from the command-line.

* New Build Invocations:
Previously, the build script simply built the Install Target. Now,
we can configure the Build Tree, build the Install Target or build 
the NSIS Package Target. For Build Tree Configuration and Install
targets, both Debug and Release are available. The NSIS Package Target
is only available for WSJT-X as MAP65 and WSPR-X do not have package
targets in their CmakeLists.txt files. Help files are provided which
list the new invocations. In the event a user inputs the wrong
the instruction set. The user can also list then by simply typing:
( build-help ) in the JTSDK-QT environment.

* Dev Guide Updates:
The JTSDK-QT section needs to be fully re-written, and many
additional features added. As TIme Allows, I'll get to this but I want
flush out any issues of the update in parallel.



git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/doc/dev-guide@4242 ab8295b8-cf94-4d9e-aec4-7959e3be5d79