JTSDK-QT add MSYS Hamlib3 build script
This script is not downloaded to JTSDK-QT, as it is an optional tool.
To use is, export the file to your MSYS $USER location:

svn export https://svn.code.sf.net/p/wsjt/wsjt/branches/doc/dev-guide/scripts/build-hamlib.sh

USAGE
./build-hamlib.sh

REQUIRMENTS
* Requires Git, Autotools and Bash to be in the MSYS path.
* Provides a link to MSYS2 that has all the VCS tools + autotools
pre-installed.
* It is set to use JTSDK-QT\QT5\Tools tool-chain to build with, which is 
required for WSJT-X and can be edited as needed.

NOTES
- Care should be taken to ensure the original MSYS tool-chain is
not in the path bash $PATH at the time of build.
- You will need to adjust the path to MSYS in the script as it's set
to my ENV which is /d/msys32/msys/bin
- Optionally, you can test parallel builds by changing the comments
lines ( approx 120 / 121 ) to:

mingw32-make -s -jx

Where "x" is the number of cores you would like to use. Bare in mind,
parallel builds on WIndows using mingw32-make is a new feature in the
tool-chain and may not always produce the expected results.


git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/doc@4406 ab8295b8-cf94-4d9e-aec4-7959e3be5d79