Initial super buildp project for WSJT-X
There is  a need for  an all  encompassing project that  builds WSJT-X
from nothing on a Linux system. This is required so the project can be
built on a  Linux build server for preparation of  a Linux installtion
package.

This project  has a single  source file which  is a CMake  script. The
script when executed  pulls the underlying project  sources from their
respective source  control system public repositories  and then builds
the application from them.

It  may also  be  used to  bootstrap  a  build on  Linux  for a  local
installation for  those that wish to  build from source on  a platform
that doesn't have a pre-built package available.

git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx-superbuild@4494 ab8295b8-cf94-4d9e-aec4-7959e3be5d79