Target to build a self-contained source tarball and build from it
On package build servers like the  Debian ones the build script is not
able to  access the Internet. This  defeats the purpose of  the WSJT-X
superbuild script  since it  relies on  the VCS  to fetch  the project
sources as part of the build.

To work  around this the  build script has been  enhanced to do  a two
phase build process.

The first  phase is a  conventional build that  runs on a  normal host
with a  normal configure  step. It  produces a  source tarball  of the
project that is  augmented with the extra sources  that would normally
be  fetched  in  the  build  process. This  first  phase  is  done  by
specifying the 'source' target when building.

The second phase  is done using the  tarball built in phase  one as an
upstream source tarball and is built  with a special option to specify
a local build.

git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx-superbuild@4567 ab8295b8-cf94-4d9e-aec4-7959e3be5d79