Add on-line KVASD install to Windows NSIS installer
To comply  with the  WSJT_X GPL  v3 license,  the closed  source KVASD
executable  must  be  delivered   separately  and  optionally  to  end
users. To facilitate  this a new install option has  been added to the
Windows  installer  to install  KVASD  by  downloading it  at  install
time. The installer  first downloads the KVASD EULA  from the Internet
and presents it  to the user for reading and  agreement.  If agreement
is  given  the  KVASD  binary  is downloaded  from  the  Internet  and
checksummed using an  MD5 hash, if that passes, it  is installed along
with WSJT-X.  An  uninstall record is also written to  ensure that, if
installed, KVASD will be un-installed if WSJT-X is un-installed.

To facilitate this enhancment to the NSIS installer the standard CPack
NSIS  installer script  template has  been overridden  with a  project
specific version.   This version also  implements the more  modern and
recomended nsDialogs mechanism for  installer custom page construction
including the install options page. See CMake/Modules/NSIS.template.in
for  details   and  CMakeCPackOptions.cmake.in   for  usage   in  this
installer.

Merged from branch wsjtx-1.4.



git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@4663 ab8295b8-cf94-4d9e-aec4-7959e3be5d79