Add rigctl to WSJT-X package as rigctl-wsjtx
The rigctl program is not used  directly by WSJT-X but it does contain
the  same  Hamlib  code  from  the   Hamlib  v3  fork  as  WSJT-X  and
rigctld-wsjtx. So it makes sense to distribute it as a testing tool or
for any user that needs it to converse with the rigctld-wsjtx server.

git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx-1.4@4774 ab8295b8-cf94-4d9e-aec4-7959e3be5d79