Don't use QSerialPortInfo::isBusy() to filter available ports
The above  member opens the  port and  that twiddles DTR/RTS  which is
unhelpful if they are connected to  PTT. Instead a primitive port name
filter  to try  and  detect unuseable  ends of  serial  port pairs  is
employed. This may  need further tests to exclude  similar devices yet
to be discovered.

git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@4994 ab8295b8-cf94-4d9e-aec4-7959e3be5d79