Tidy up resource generation with a CMake function
This  change  generates  all  the resources  using  a  CMake  function
add_resources()  and  substitution  into  a template  .qrc  file.  The
location of resource files in the  source tree is no longer related to
the path  of the resources  in the embedded executable  resources file
system.  Instead the roesources  are embedded with predetermined paths
e.g.  :/, :/Palettes/ and, :/samples/ .

git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@5131 ab8295b8-cf94-4d9e-aec4-7959e3be5d79