Band activity & Rx frequency widget enhancements
Ensure that  these windows have  their viewport scrolled fully  to the
left after deocdes and font changes.

Add the band to the separator line but only if we are certain that the
decodes below it are on that band.

Thanks to  Ton PA0TBR for the  implementation and testing of  the core
features of this change.

git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@5207 ab8295b8-cf94-4d9e-aec4-7959e3be5d79