First attempt at adding the WSJT-X user guide to the CMake build
These documentation source files are not  the one true version, just a
copy for testing purposes. DO NOT EDIT THESE FILES.

To use this  on Windows you will need a  working asciidoc installation
and  the  path  to  it  must be  included  in  your  CMAKE_PREFIX_PATH
(probably via a  local CMake tool chain file). At  the time of writing
the official  asciidoc package does  not work on Windows.   The latest
development  master does  however  work,  it can  be  downloaded as  a
snapshot ZIP archive from here:

  https://github.com/asciidoc/asciidoc/archive/master.zip

git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@5316 ab8295b8-cf94-4d9e-aec4-7959e3be5d79