Make UDP message handling version mismatch tolerant
By  providing safe  defaults for  expected input  fields that  are not
present in messages from older version  senders, it is now possible to
mix  versions of  clients and  servers.   This relys  on fields  never
changing meaning or content, new fields may be safely added to the end
of existing messages.

git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx-1.5@5337 ab8295b8-cf94-4d9e-aec4-7959e3be5d79