Rig frequency calibration
The   settings   are  a   linear   calibration   correction  for   the
rig. Calibration constants are supplied as an intercept error in Hertz
and a slope error in parts per million.

The correction is  applied only to the final frequencies  going to and
from  the radio.   This means  that  any transverter  offsets are  not
corrected. This is  appropriate as each transverter will  have its own
error factors.  Transvertor and  other individual band specific errors
can only be corrected by adjusting the offset for the band.

Thanks  to Mike  W9MDB for  an  initial implementation  of Joe  K1JT's
suggestion. See also the WSPR documentation for details of calculating
the required calibration constants.

git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@5396 ab8295b8-cf94-4d9e-aec4-7959e3be5d79