Reset eliding of duplicate UDP messages if a Replay message is received
UDP message servers will not receive  a Status until some status field
changes if  they join a network  with existing clients. To  avoid this
receipt  of a  Replay message,  which should  be the  first message  a
server sends  after negotiating the  schema, now cuases  the duplicate
message check  to be  disabled for  the next  message. This  should be
sufficient for servers to get into sync.

git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@6459 ab8295b8-cf94-4d9e-aec4-7959e3be5d79