Improve the "translatability" of some of the User Guide links
To  aid translators  some of  the links  in the  doc/common/links.adoc
file, which is shared by all translations, have been changed to remove
generic  link  text  like  "here"   with  the  titles  of  the  linked
references.   Since  these  are  in the  language  of  the  references
documents there is no need to translate these.

git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@6647 ab8295b8-cf94-4d9e-aec4-7959e3be5d79