Ignore OmniRig frequency changes while transmitting on single VFO reporting rigs
Many rigs  with A/B VFO  arrangements and single VFO  reporting report
the split  Tx VFO frequency  when transmitting. This  gets complicated
when  using VOX  as we  do not  know  exactly when  the rig  is in  Tx
mode. This change ignores frequency changes from OmniRig for this sort
of rig while  transmitting to try and improve rig  tracking. It is not
perfect but seems better than before.

git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@6688 ab8295b8-cf94-4d9e-aec4-7959e3be5d79