FMT - now builds all four binaries
* builds fmtest, fmtave, fcal and fmeasure
* libfmt. needs to be reviewed for accuracy

I performed a test run using the values in examples/gocal. The final
corrections matched closely with my current setting obtained with the tools
build in WSPR. Still needed more validation.

The installation is made with the project directory ./install. The install
folder is removed with: make distclean

TO TEST
-------
* checkout from svn
* ./autogen.sh
* make
* make install
* cd ./install
* python -O fmtparams.py
* Fill out and save Station Parameters
* Edit gocal frequencies
* ./gocal
* fmtave fmt.all
* fcal fmtave.out

Verify data settings with WSJT-X or another application that takes offsets.



git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/fmt@6819 ab8295b8-cf94-4d9e-aec4-7959e3be5d79