Fix mode setting with OmniRig and allow rig type changes
The OmniRig interface now  tries to set the Tx VFO  mode when in split
operating. OmniRig already  does this for some types of  rigs where it
doesn't glitch  the Rx, this  change also  deals with those  that will
glitch the Rx e.g. where VFO swap is needed.

This change also  tries to set the  VFO to A where the  rig only split
Tx's on the other.

OmniRig allows the rig  type to be changed on the  fly, this change no
longer treats that  as a rig control error and  tries to continue with
the new  rig type. Also better  handling of OmniRig reporting  the rig
offline now allows up to five seconds for OmniRig to return the rig to
online.

git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@7032 ab8295b8-cf94-4d9e-aec4-7959e3be5d79