Do not wait for PTT report from DX Lab Suite Commander when using direct DTR or RTS PTT
The recent  change to delay  Tx audio  until Commander reports  PTT is
really asserted should not be done when a direct RTS or DTR PTT method
is  being used.   This does  leave a  remote chance  of ALC  overshoot
issues from audio starting too early so  it is down to the user to set
an   appropriate  "Settings->Advanced->Tx Delay"   to   ensure    safe
operating.

git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@7774 ab8295b8-cf94-4d9e-aec4-7959e3be5d79