Minimum JT9 decode offset maintained per band
This  change allows  allocations for  JT9+JT65  dual mode  to be  fine
tuned. This  in turn will hopefully  allow some space for  the new FT8
mode to be  shoehorned in on bands  with little or no  free space. For
example on  top band users could  set the lowest JT9  decode offset at
1000 Hz and  set the FT8 default working frequency  at 1.839 MHz. Thus
allowing 1000 Hz for JT65, 1000 Hz for JT9 and 1000 Hz for FT8.

git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@7809 ab8295b8-cf94-4d9e-aec4-7959e3be5d79