Make interaction with rigs using Hamlib rigctld as a proxy work better
The Hamlib rigctld does  not allow the local end to  see teh exact rig
capabilities,  this means  that  some tests  for  capabilities do  not
provide the correct status. This  change attempts to work around these
limitations by testing the wanted capabilities by trying to use them.

Thanks   for   Mike,  W9MDB,   for   the   initial  work   and   patch
contribution. This change is loosely based on his patch.

git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@7818 ab8295b8-cf94-4d9e-aec4-7959e3be5d79