Improved CAT via HRD for rigs with VFO B Tx but limited VFO B Rx
Rigs  like  the  Yaesu  FT-991  can  split but  VFO  B  is  not  fully
independent so there is  no way to switch to VFO B,  these rigs do not
appear to have the possibility of  different modulation modes on VFO A
and VFO B so  we don't have to worry about the Tx  VFO mode when using
split.

Merged from wsjtx-1.8 branch.



git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@8307 ab8295b8-cf94-4d9e-aec4-7959e3be5d79