Add blocking mechanism for UDP message destinations and enforce for pskreporter addresses
Avoid fragmentation of pskreporter destined datagrams

Keep  datagram size  below  a  reasonable IPv4  MTU  size of  protocol
headers + 1400.

Thanks to Philip,  N1DQ, for the suggestion as with  FT8 mode datagram
payload sizes to Pskrepotrer are sometimes as big a 8kBytes.

Correct a factor of two error in Pskreporter UDP message sizing

Data is built  as a hexadecmal string,  so is twice the  length of the
message bytes it is encoded as.

Merged from trunk.



git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx-1.8@8345 ab8295b8-cf94-4d9e-aec4-7959e3be5d79