Allow partial processing of UDP Reply messages for non-CQ/QRZ messages
This allows external programs to set up a QSO, for example to tail-end
a  QSO,  but unlike  double-clicking  a  non-CQ/QRZ decode  in  WSJT-X
directly it will  not enable auto Tx. This last  constraint is to make
fully automating WSJT-X using UDP messages impossible.

It is expected that programs  like JTAlert will, without change, allow
users to set  up a QSO with  any station from the  JTAlert UI although
for  non-CQ/QRZ decodes  a  QSO can  only be  started  by pushing  the
"Enable Tx" button in WSJT-X.

Thanks to Mike, W9MDB, for the patch that this change is based on, the
implementation  committed is  considerably simpler  than his  proposed
change.

git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@8574 ab8295b8-cf94-4d9e-aec4-7959e3be5d79